baru/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pandas==2.2.2
joblib==1.4.2
scikit-learn==1.6.1
catboost==1.2.8
python-dotenv==1.0.1
orjson==3.9.12
psycopg2-binary==2.9.9
apscheduler==3.10.4
tensorflow-cpu>=2.16.1
keras>=3.0.0
pyarrow
numpy==2.0.2
